net/oic; move transport specific declarations to their respective header files.
