+++ net/oic/src/port/mynewt/ble_adaptor.c
@@ -31,8 +31,9 @@
 #include "oic/oc_ri.h"
 #include "messaging/coap/coap.h"
 #include "api/oc_buffer.h"
-#include "port/oc_connectivity.h"
+#include "oic/port/oc_connectivity.h"
 #include "adaptor.h"
+#include "oic/port/mynewt/ble.h"
 #include "host/ble_hs.h"
 #include "services/gap/ble_svc_gap.h"
 #include "services/gatt/ble_svc_gatt.h"
