+++ net/oic/src/port/mynewt/lora_adaptor.c
@@ -33,11 +33,12 @@
 
 #include <node/lora.h>
 
-#include "port/oc_connectivity.h"
+#include "oic/port/oc_connectivity.h"
 #include "oic/oc_log.h"
 #include "api/oc_buffer.h"
 #include "adaptor.h"
 #include "oic/port/mynewt/transport.h"
+#include "oic/port/mynewt/lora.h"
 
 #ifdef OC_SECURITY
 #error This implementation does not yet support security
